improve health code
