+++ project/main/business/business_helper.py
@@ -9,55 +9,19 @@ import string
 
 import project.database.utils as utils
 import project.main.util_helper as util_helper
+import project.main.same_function_helper as same_helper
 
 from project.database.models import GasSensor, Qhawax, EcaNoise, QhawaxInstallationHistory, Company, AirQualityMeasurement, ProcessedMeasurement, ValidProcessedMeasurement, Bitacora
-
 
 var_gases=['CO','H2S','NO','NO2','O3','SO2']
 
 session = db.session
 
 
-def getQhawaxID(qhawax_name): 
-    """
-    Get qHAWAX ID base on qHAWAX name
-
-    :type qhawax_name: string
-    :param qhawax_name: qHAWAX name
-
-    """
-    qhawax_list = session.query(Qhawax.id).filter_by(name=qhawax_name).all()
-    if(qhawax_list == []):
-        raise TypeError("The qHAWAX name could not be found")
-    return session.query(Qhawax.id).filter_by(name=qhawax_name).one()[0]
-
-
 def getInstallationIdBaseName(qhawax_name):
-    qhawax_id = getQhawaxID(qhawax_name)
-    installation_id = getInstallationId(qhawax_id)
+    qhawax_id = same_helper.getQhawaxID(qhawax_name)
+    installation_id = same_helper.getInstallationId(qhawax_id)
     return installation_id
-
-def getInstallationId(qhawax_id):
-    installation_id= session.query(QhawaxInstallationHistory.id).filter_by(qhawax_id=qhawax_id). \
-                                    filter(QhawaxInstallationHistory.end_date == None). \
-                                    order_by(QhawaxInstallationHistory.instalation_date.desc()).all()
-    if(installation_id == []):
-        return None
-
-    return session.query(QhawaxInstallationHistory.id).filter_by(qhawax_id=qhawax_id). \
-                                    filter(QhawaxInstallationHistory.end_date == None). \
-                                    order_by(QhawaxInstallationHistory.instalation_date.desc()).first()[0]
-
-def getQhawaxName(qhawax_id):
-    """
-    Get qHAWAX Name
-
-    :type qhawax_id: integer
-    :param qhawax_id: qHAWAX ID
-
-    """
-    return session.query(Qhawax.name).filter(Qhawax.id == qhawax_id).one()   
-
 
 def getMainIncaQhawaxTable(qhawax_id):
     """
@@ -332,7 +296,7 @@ def saveStatusOn(qhawax_name):
 def saveTurnOnLastTime(qhawax_name):
     qhawax_id, mode = session.query(Qhawax.id, Qhawax.mode).filter_by(name=qhawax_name).first()
     if(mode=='Cliente'):
-        installation_id = getInstallationId(qhawax_id)
+        installation_id = same_helper.getInstallationId(qhawax_id)
         if(installation_id!=None):
             now = datetime.datetime.now(dateutil.tz.tzutc())
             session.query(QhawaxInstallationHistory).filter_by(id=installation_id).update(values={'last_time_physically_turn_on_zone': now.replace(tzinfo=None)})
